refactor: editor extensions
