+++ packages/editor/src/core/extensions/read-only-extensions.ts
@@ -4,7 +4,6 @@ import TaskItem from "@tiptap/extension-task-item";
 import TaskList from "@tiptap/extension-task-list";
 import TextStyle from "@tiptap/extension-text-style";
 import TiptapUnderline from "@tiptap/extension-underline";
-import StarterKit from "@tiptap/starter-kit";
 import { Markdown } from "tiptap-markdown";
 // extensions
 import {
@@ -25,14 +24,13 @@ import {
   UtilityExtension,
   ImageExtension,
 } from "@/extensions";
-// helpers
-import { isValidHttpUrl } from "@/helpers/common";
 // plane editor extensions
 import { CoreReadOnlyEditorAdditionalExtensions } from "@/plane-editor/extensions";
 // types
 import type { IReadOnlyEditorProps } from "@/types";
 // local imports
 import { CustomImageExtension } from "./custom-image/extension";
+import { CustomStarterKitExtension } from "./starter-kit";
 
 type Props = Pick<IReadOnlyEditorProps, "disabledExtensions" | "flaggedExtensions" | "fileHandler" | "mentionHandler">;
 
@@ -40,56 +38,12 @@ export const CoreReadOnlyEditorExtensions = (props: Props): Extensions => {
   const { disabledExtensions, fileHandler, flaggedExtensions, mentionHandler } = props;
 
   const extensions = [
-    StarterKit.configure({
-      bulletList: {
-        HTMLAttributes: {
-          class: "list-disc pl-7 space-y-[--list-spacing-y]",
-        },
-      },
-      orderedList: {
-        HTMLAttributes: {
-          class: "list-decimal pl-7 space-y-[--list-spacing-y]",
-        },
-      },
-      listItem: {
-        HTMLAttributes: {
-          class: "not-prose space-y-2",
-        },
-      },
-      code: false,
-      codeBlock: false,
-      horizontalRule: false,
-      blockquote: false,
-      paragraph: {
-        HTMLAttributes: {
-          class: "editor-paragraph-block",
-        },
-      },
-      heading: {
-        HTMLAttributes: {
-          class: "editor-heading-block",
-        },
-      },
-      dropcursor: false,
-      gapcursor: false,
+    CustomStarterKitExtension({
+      enableHistory: false,
     }),
     CustomQuoteExtension,
-    CustomHorizontalRule.configure({
-      HTMLAttributes: {
-        class: "py-4 border-custom-border-400",
-      },
-    }),
-    CustomLinkExtension.configure({
-      openOnClick: true,
-      autolink: true,
-      linkOnPaste: true,
-      protocols: ["http", "https"],
-      validate: (url: string) => isValidHttpUrl(url).isValid,
-      HTMLAttributes: {
-        class:
-          "text-custom-primary-300 underline underline-offset-[3px] hover:text-custom-primary-500 transition-colors cursor-pointer",
-      },
-    }),
+    CustomHorizontalRule,
+    CustomLinkExtension,
     CustomTypographyExtension,
     TiptapUnderline,
     TextStyle,
@@ -104,11 +58,7 @@ export const CoreReadOnlyEditorExtensions = (props: Props): Extensions => {
       },
       nested: true,
     }),
-    CustomCodeBlockExtension.configure({
-      HTMLAttributes: {
-        class: "",
-      },
-    }),
+    CustomCodeBlockExtension,
     CustomCodeInlineExtension,
     Markdown.configure({
       html: true,
